Fixed test suite bug caused by running root in my user home directory.
Test previously just used TMPDIR, but running fetchware clean there in an
automated fashion could interfere with normal TMPDIR use, so it was foolishly
change to use cwd(), which blows up when root due to security checking code.

I fixed it by change it to use a "temporary tempdir": basically just a tempdir
that more tempdirs are made in. This directory is then fetchware cleaned without
messing with main TMPDIR or using cwd() which blows up when root.