Fixed previous botched fix in create_fetchware_package().
In a previous commit, cb7b0b3f4e672cf1b26b7f71e075e15c4d023df8, I fixed a bug
where when no_install was set a fetchware package would still be created and
copied to your fetchware database path. It would not actually be installed, but
it could be if a fetchware upgrade-all is run, which would technically "upgrade"
this package, but this would really just install it.

For some reason another change in create_fetchware_package() was also created in
the commit. I don't know why or what it's really supposed to fix, especially
since drop_privs() does not actually work the way this patch foolishly assumes
it works. drop_privs chdir()s to the directory, then it forks this way both
parent and child are in the same directory making everything work easier with
fewer differences between parent and child.

This broken patch has been reverted, which fixes create_fetchware_package() for
root users.