Fixed root drop_privs() bug cause by ignorance of extra temporary directory.
cmd_look() when catfile()ing the look directory to tell the user what directory
to cd too it use $temp_dir, but $temp_dir does not account for the extra
directory drop_privs() creates. So, I switched it to use $package_path, which
does have that extra directory in it fixing this bug.