Fix failing tests due to botched 'uninitialized' warning removal.
Some complicated if statements that I could make any sense of had botched their
'uninitialized' warning removals. The biggest problem was my foolish assumption
that both if statements tested for the same thing, which of course they didn't.
Their crazy conditional was in fact reversed.

Fixed both with not one but two ternary operators inside an already complicated
if statement. Comments were needed to help say what the hell they test for. I
tried may other things, but this is the only one to work. I considered adding
another if statement inside the first one for it, but that's less complicated
but just a confusing, so not much better.