Breaking Change: umask 0022 now enforced by fetchware for security reasons.
Prior to this commit fetchware was happy using whatever the user's umask was.
But some CPAN Tester's have umasks of 0002, which cause fetchware to create
files that are group writable. These files are occasionally read back into
fetchware to test certain things, and these files, then fail fetchware's
safe_open() security checks.

An example CPAN Tester FAIL is:
http://www.cpantesters.org/cpan/report/967600c8-8a03-11e6-ab41-c893a58a4b8c

In that CPAN Fail, the perms of /tmp are almost certainly a safe 01777, and the
File::Temp created tempdir is most likely 0700 since File::Temp seems to ignore
umask. So, that leaves the fetchware package itself, and under a umask of 0002
it would fail safe_open()'s checks.