Added umask change to all test suite fales so testing is same as fetchware.
I considered adding the umask change to Test::Fetchware and
App::Fetchware::Util, but those are libraries *not* programs, and programs
should dictate what the umask is *not* a library.

So, instead, I added the umask change to the entire test suite. Now the test
suite will operate under the same umask conditions as fetchware itself.